+++ aizynthfinder/utils/route_clustering/__init__.py
@@ -1,4 +1,0 @@
-""" Package for clustering and distance calculations
-"""
-from .clustering import ClusteringHelper  # noqa
-from .distances import TreeContent, ReactionTreeWrapper  # noqa